use device selector
